# docs/public/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UsageEntry": {
        "description": "A single usage log entry.",
        "properties": {
          "api_key_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Id"
          },
          "completion_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Tokens"
          },
          "cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "prompt_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Tokens"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "additionalProperties": true,
            "title": "Tags",
            "type": "object"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "user_id",
          "api_key_id",
          "project_id",
          "timestamp",
          "model",
          "provider",
          "endpoint",
          "prompt_tokens",
          "completion_tokens",
          "total_tokens",
          "cost",
          "status",
          "error_message",
          "tags"
        ],
        "title": "UsageEntry",
        "type": "object"
      },
      "UsageLogResponse": {
        "description": "Response model for usage log.",
        "properties": {
          "api_key_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Id"
          },
          "completion_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Tokens"
          },
          "cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "prompt_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Tokens"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "additionalProperties": true,
            "title": "Tags",
            "type": "object"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "user_id",
          "api_key_id",
          "project_id",
          "timestamp",
          "model",
          "provider",
          "endpoint",
          "prompt_tokens",
          "completion_tokens",
          "total_tokens",
          "cost",
          "status",
          "error_message",
          "tags"
        ],
        "title": "UsageLogResponse",
        "type": "object"
      },
      "UsageSummaryBucket": {
        "description": "Aggregated usage metrics for one grouping key.",
        "properties": {
          "completion_tokens": {
            "title": "Completion Tokens",
            "type": "integer"
          },
          "cost": {
            "title": "Cost",
            "type": "number"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "prompt_tokens": {
            "title": "Prompt Tokens",
            "type": "integer"
          },
          "success_count": {
            "title": "Success Count",
            "type": "integer"
          },
          "total_tokens": {
            "title": "Total Tokens",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "count",
          "success_count",
          "error_count",
          "prompt_tokens",
          "completion_tokens",
          "total_tokens",
          "cost"
        ],
        "title": "UsageSummaryBucket",
        "type": "object"
      },
      "UsageSummaryResponse": {
        "description": "Aggregated usage metrics.",
        "properties": {
          "by_endpoint": {
            "items": {
              "$ref": "#/components/schemas/UsageSummaryBucket"
            },
            "title": "By Endpoint",
            "type": "array"
          },
          "by_model": {
            "items": {
              "$ref": "#/components/schemas/UsageSummaryBucket"
            },
            "title": "By Model",
            "type": "array"
          },
          "by_project": {
            "items": {
              "$ref": "#/components/schemas/UsageSummaryBucket"
            },
            "title": "By Project",
            "type": "array"
          },
          "by_provider": {
            "items": {
              "$ref": "#/components/schemas/UsageSummaryBucket"
            },
            "title": "By Provider",
            "type": "array"
          },
          "by_status": {
            "items": {
              "$ref": "#/components/schemas/UsageSummaryBucket"
            },
            "title": "By Status",
            "type": "array"
          },
          "by_tag": {
            "items": {
              "$ref": "#/components/schemas/UsageSummaryBucket"
            },
            "title": "By Tag",
            "type": "array"
          },
          "by_user": {
            "items": {
              "$ref": "#/components/schemas/UsageSummaryBucket"
            },
            "title": "By User",
            "type": "array"
          },
          "completion_tokens": {
            "title": "Completion Tokens",
            "type": "integer"
          },
          "cost": {
            "title": "Cost",
            "type": "number"
          },
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "prompt_tokens": {
            "title": "Prompt Tokens",
            "type": "integer"
          },
          "success_count": {
            "title": "Success Count",
            "type": "integer"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          },
          "total_tokens": {
            "title": "Total Tokens",
            "type": "integer"
          }
        },
        "required": [
          "total_count",
          "success_count",
          "error_count",
          "prompt_tokens",
          "completion_tokens",
          "total_tokens",
          "cost",
          "by_project",
          "by_user",
          "by_model",
          "by_provider",
          "by_endpoint",
          "by_status",
          "by_tag"
        ],
        "title": "UsageSummaryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "A clean FastAPI gateway for otari with API key management",
    "title": "otari-gateway",
    "version": "0.0.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/usage": {
      "get": {
        "description": "List usage logs ordered by timestamp (most recent first).\n\nSupports optional filters for time range, user, project, and usage tags.\nPaginated via skip/limit. Timestamps accept either ISO 8601 strings or\nUnix epoch seconds (numeric).",
        "operationId": "list_usage_v1_usage_get",
        "parameters": [
          {
            "description": "Return logs with timestamp >= start_date (ISO 8601 or Unix epoch seconds)",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return logs with timestamp >= start_date (ISO 8601 or Unix epoch seconds)",
              "title": "Start Date"
            }
          },
          {
            "description": "Return logs with timestamp < end_date (ISO 8601 or Unix epoch seconds)",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return logs with timestamp < end_date (ISO 8601 or Unix epoch seconds)",
              "title": "End Date"
            }
          },
          {
            "description": "Filter to a single user",
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a single user",
              "title": "User Id"
            }
          },
          {
            "description": "Filter to a gateway project",
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a gateway project",
              "title": "Project Id"
            }
          },
          {
            "description": "Filter to logs containing this usage tag key",
            "in": "query",
            "name": "tag_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to logs containing this usage tag key",
              "title": "Tag Key"
            }
          },
          {
            "description": "When tag_key is set, filter to logs whose tag value string matches this value",
            "in": "query",
            "name": "tag_value",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "When tag_key is set, filter to logs whose tag value string matches this value",
              "title": "Tag Value"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsageEntry"
                  },
                  "title": "Response List Usage V1 Usage Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Usage",
        "tags": [
          "usage"
        ]
      }
    },
    "/v1/usage/summary": {
      "get": {
        "description": "Summarize usage logs with optional project and tag filters.",
        "operationId": "summarize_usage_v1_usage_summary_get",
        "parameters": [
          {
            "description": "Return logs with timestamp >= start_date (ISO 8601 or Unix epoch seconds)",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return logs with timestamp >= start_date (ISO 8601 or Unix epoch seconds)",
              "title": "Start Date"
            }
          },
          {
            "description": "Return logs with timestamp < end_date (ISO 8601 or Unix epoch seconds)",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return logs with timestamp < end_date (ISO 8601 or Unix epoch seconds)",
              "title": "End Date"
            }
          },
          {
            "description": "Filter to a single user",
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a single user",
              "title": "User Id"
            }
          },
          {
            "description": "Filter to a gateway project",
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to a gateway project",
              "title": "Project Id"
            }
          },
          {
            "description": "Filter to logs containing this usage tag key",
            "in": "query",
            "name": "tag_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to logs containing this usage tag key",
              "title": "Tag Key"
            }
          },
          {
            "description": "When tag_key is set, filter to logs whose tag value string matches this value",
            "in": "query",
            "name": "tag_value",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "When tag_key is set, filter to logs whose tag value string matches this value",
              "title": "Tag Value"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "maximum": 10000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Summarize Usage",
        "tags": [
          "usage"
        ]
      }
    },
...
    "/v1/users/{user_id}/usage": {
      "get": {
        "description": "Get usage history for a specific user.",
        "operationId": "get_user_usage_v1_users__user_id__usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsageLogResponse"
                  },
                  "title": "Response Get User Usage V1 Users  User Id  Usage Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get User Usage",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}